Crear Inmueble con Foto

+++ app/src/main/java/com/example/inmobiliaria/model/Inmueble.java
@@ -106,7 +106,7 @@ public class Inmueble implements Serializable {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
